+++ setup.cfg
@@ -15,8 +15,11 @@ classifiers =
     Development Status :: 5 - Production/Stable
 
 [options]
+package_dir=
+    =src
 packages = find:
 python_requires = >=3.8
 include_package_data = False
-install_requires =
+install_requires = [
     "pygame>=2.1.2"
+]
