+++ src/pygame_gridcalculator/gridcalculator.py
