+++ src/pygame_gridcalculator/shapefactory.py
